Fix tests to work with PDU event trait
