+++ src/lib.rs
@@ -19,7 +19,7 @@ mod state_store;
 
 pub use error::{Error, Result};
 pub use event_auth::{auth_check, auth_types_for_event};
-pub use state_event::{Event, Requester};
+pub use state_event::Event;
 pub use state_store::StateStore;
 
 // We want to yield to the reactor occasionally during state res when dealing
@@ -154,7 +154,7 @@ impl StateResolution {
             .unwrap();
 
         // update event_map to include the fetched events
-        event_map.extend(events.into_iter().map(|ev| (ev.event_id(), ev)));
+        event_map.extend(events.into_iter().map(|ev| (ev.event_id().clone(), ev)));
         // at this point our event_map == store there should be no missing events
 
         tracing::debug!("event map size: {}", event_map.len());
@@ -383,7 +383,7 @@ impl StateResolution {
             // This return value is the key used for sorting events,
             // events are then sorted by power level, time,
             // and lexically by event_id.
-            (-*pl, ev.origin_server_ts(), ev.event_id())
+            (-*pl, ev.origin_server_ts(), ev.event_id().clone())
         })
     }
