+++ src/state_store.rs
@@ -1,9 +1,6 @@
 use std::{collections::BTreeSet, sync::Arc};
 
-use ruma::{
-    events::pdu::ServerPdu,
-    identifiers::{EventId, RoomId},
-};
+use ruma::identifiers::{EventId, RoomId};
 
 use crate::{Event, Result};
