+++ tests/event_auth.rs
@@ -6,11 +6,11 @@ use state_res::{
         // auth_check, auth_types_for_event, can_federate, check_power_levels, check_redaction,
         valid_membership_change,
     },
-    Requester, StateMap
+    StateMap
 };
 
 mod utils;
-use utils::{alice, charlie, event_id, member_content_ban, room_id, INITIAL_EVENTS};
+use utils::{alice, charlie, event_id, member_content_ban, room_id, to_pdu_event, INITIAL_EVENTS};
 
 #[test]
 fn test_ban_pass() {
@@ -18,23 +18,25 @@ fn test_ban_pass() {
 
     let prev = events
         .values()
-        .find(|ev| ev.event_id.as_str().contains("IMC"))
+        .find(|ev| ev.event_id().as_str().contains("IMC"))
         .map(Arc::clone);
 
     let auth_events = events
         .values()
-        .map(|ev| ((ev.kind.clone(), ev.state_key.clone()), Arc::clone(ev)))
+        .map(|ev| ((ev.kind(), ev.state_key()), Arc::clone(ev)))
         .collect::<StateMap<_>>();
 
-    let requester = Requester {
-        prev_event_ids: vec![event_id("IMC")],
-        room_id: &room_id(),
-        content: &member_content_ban(),
-        state_key: Some(charlie().to_string()),
-        sender: &alice(),
-    };
+    let requester = to_pdu_event(
+        "HELLO",
+        alice(),
+        ruma::events::EventType::RoomMember,
+        Some(charlie().as_str()),
+        member_content_ban(),
+        &vec![],
+        &vec![event_id("IMC")],
+    );
 
-    assert!(valid_membership_change(requester, prev, None, &auth_events).unwrap())
+    assert!(valid_membership_change(&requester, prev, None, &auth_events).unwrap())
 }
 
 #[test]
@@ -43,21 +45,23 @@ fn test_ban_fail() {
 
     let prev = events
         .values()
-        .find(|ev| ev.event_id.as_str().contains("IMC"))
+        .find(|ev| ev.event_id().as_str().contains("IMC"))
         .map(Arc::clone);
 
     let auth_events = events
         .values()
-        .map(|ev| ((ev.kind.clone(), ev.state_key.clone()), Arc::clone(ev)))
+        .map(|ev| ((ev.kind(), ev.state_key()), Arc::clone(ev)))
         .collect::<StateMap<_>>();
 
-    let requester = Requester {
-        prev_event_ids: vec![event_id("IMC")],
-        room_id: &room_id(),
-        content: &member_content_ban(),
-        state_key: Some(alice().to_string()),
-        sender: &charlie(),
-    };
+    let requester = to_pdu_event(
+        "HELLO",
+        charlie(),
+        ruma::events::EventType::RoomMember,
+        Some(alice().as_str()),
+        member_content_ban(),
+        &vec![],
+        &vec![event_id("IMC")],
+    );
 
-    assert!(!valid_membership_change(requester, prev, None, &auth_events).unwrap())
+    assert!(!valid_membership_change(&requester, prev, None, &auth_events).unwrap())
 }
